+++ doc/release_notes.rst
@@ -1,6 +1,8 @@
 Release notes
 =============
 
+* Quick polarization analysis routine in place.
+
 
 *ximpol (0.22.1) - Mon, 08 Feb 2016 15:11:38 -0800*